GraphGenerator/requirements.txt
matplotlib~=3.7.1
numpy~=1.24.3
nibabel~=5.1.0
pandas~=2.0.1
Pillow~=9.5.0